change hooks to allow review assignment display by section
