+++ src/utils/hooks/useLoadApplicationNEW.ts
@@ -152,7 +152,7 @@ const useLoadApplication = ({ serialNumber, networkFetch }: UseGetApplicationPro
     }
     evaluate(application.template?.submissionMessage || '', evaluatorParams).then(
       (submissionMessage: any) => {
-        setFullStructure({
+        let newStructure: FullStructure = {
           info: { ...applicationDetails, submissionMessage },
           stages: templateStages.map((stage) => ({
             number: stage.number as number,
@@ -160,7 +160,11 @@ const useLoadApplication = ({ serialNumber, networkFetch }: UseGetApplicationPro
             description: stage.description ? stage.description : undefined,
           })),
           sections: buildSectionsStructure({ sections, baseElements }),
-        })
+        }
+        // This is usefull for generating progress (maybe also usefull downstream, but dont' want too refactor too much at this stage)
+        newStructure = addSortedSectionsAndPages(newStructure)
+
+        setFullStructure(newStructure)
         setIsLoading(false)
       }
     )
@@ -172,6 +176,19 @@ const useLoadApplication = ({ serialNumber, networkFetch }: UseGetApplicationPro
     structure,
     template,
   }
+}
+
+const addSortedSectionsAndPages = (newStructure: FullStructure): FullStructure => {
+  const sortedSections = Object.values(newStructure.sections).sort(
+    (sectionOne, sectionTwo) => sectionOne.details.index - sectionTwo.details.index
+  )
+  const sortedPages = sortedSections
+    .map((section) =>
+      Object.values(section.pages).sort((pageOne, pageTwo) => pageOne.number - pageTwo.number)
+    )
+    .flat()
+
+  return { ...newStructure, sortedPages, sortedSections }
 }
 
 export default useLoadApplication
